Generalize argument functions to allow using for fields
This patch slightly modifies the functions in arg.h to allow using
for field storage as well. It does a general rename of ..._g_arg_...
to ..._g_argument_... to make it clear that what's being referred
to is GArgument not GIArgInfo.

gjs_value_to_g_arg_with_type_info() is exposed and renamed to
gjs_value_to_g_argument and the boolean argument is changed to an
enum (should improve the error messages a bit for existing uses.)

gjs_value_to_g_arg is renamed to gjs_value_to_arg() since it *is*
specific to GIArgInfo.

svn path=/trunk/; revision=111