Improve support for machine-dependent integer types and arrays of integers.
We use a centralized function to convert all the machine-dependent integer
types into explicitly-sized machine-independent types.  This code should
constant-fold nicely under optimization.  We also remove some dead code
involving type tags.  We translate GType as an appropriately sized integer,
to allow invocation of generic functions which take a GType parameter.

We implement support for marshalling arrays of integer types from
javascript to C code.  Note that we don't check ranges here, we just use
truncation, so it is the caller's responsibility to ensure that the input
to a GI function is of appropriate magnitude.  Test cases against the
'Everything' module show how this works.