Avoid type-punning errors under optimization.
This fixes glibc's "dereferencing type-punned pointer will break
strict-aliasing rules" warning, because we're not passing a "type-punned"
pointer into a union type into a function, but instead passing in a pointer
to a "real" char * variable, and then assigning it to the union type as a
separate step.

In the second case, we pass in a 'real' void *, and cast to 'char **' in
a separate step to avoid the warning.