Bug 580948 - Add DBus support
This patch merges in code from Litl implementing a DBus module.

At a high level there are four pieces.  First, the "gjs-dbus"
library is a C support library for DBus.
Second, the modules/dbus*.[ch] implement parts of the JavaScript
API in C.  Third, the modules/dbus.js file fills out the rest
of the JavaScript API and implementation.  Fourth, there are
tests and examples.