Fix ref count of (transfer full) GObject arguments.
We need to ref the gobject when creating an in-argument so to match the
unref in g_arg_release_internal (for transfer none).  When using
(transfer full), the additional ref is the ref we are going to transfer
to the native code; our original ref is still held by the JavaScript wrapper
until the JS wrapper is garbage collected.