autoconf-2.64 compat: Don't use $(builddir) or $(abs_top_builddir)
$(builddir) is not a standard automake variable. With autoconf < 2.64
it ends up getting set in every Makefile.in to '.' (because autoconf
defines it), but that is no longer the case for 2.64.

Since $(builddir) was always '.', just use that instead.

The $(abs_top_srcdir) $(abs_top_builddir) variables are also artifacts
of old autoconf rather than something intentional in automake. To
stop using those:

 - Set the paths in TESTS_ENVIRONMENT to just be relative to
   the top build directory - we never change from there when
   running tests.
 - Make gjs-unit determine the paths to various directories by
   determining them relative to the binary.

http://bugzilla.gnome.org/show_bug.cgi?id=591897