move gjs_error_reporter into a C++ file jsapi-private.cpp
The issue is that in latest SpiderMonkey, the headers are C++.
jsapi-util.c included jscntxt.h (a C++ header) which is used
only to implement gjs_error_reporter. jsapi-private.cpp will
now contain any usage of private JS APIs that require the
C++ headers.

jsapi-private.h was added as well, but is empty; this is meant
to export APIs for use by jsapi-util.c ... currently none.