gjs_define_object_class(): generate correct object hierarchy
When we are creating a new JS class for a GObject class and look up the JS
class for the parent class, we  must make sure that we use a consistent GType
and GITypeInfo. If the GITypeInfo and the GType are mismatched, then the parent
class might be incorrectly registered. To fix this we remove the 'info'
argument from gjs_define_object_class() and always just pass in a GType.

The unused functions gjs_lookup_object_constructor() and
gjs_lookup_object_class() are removed.
https://bugzilla.gnome.org/show_bug.cgi?id=602864