[dbus] Add getCurrentMessageContext, remove old _dbus_sender
The dbus binding had inconsistent and pretty broken support for
retrieving the sender of a DBusMessage.  The function
gjs_js_add_dbus_props, when passed an object, added a _dbus_sender
property.  However, in many cases it won't be passed an object,
such as a service method implementation.

There are a number of cases here; the primary ones that we need
to support are service methods, signals, and asynchronous callbacks.

This patch regresses one case, which is that previously if you
did a synchronous call that *also* returned multiple values, you
could access _dbus_sender.  This patch has no support for synchronous
calls, however there is a workaround - convert them to asynchronous.

https://bugzilla.gnome.org/show_bug.cgi?id=609121