test: handle multiple absolute paths when building filename
When GJS_TOP_SRCDIR is an absolute path g_build_filename constructs the
paths in a way we're not expecting, e.g. given path elements

    ["/top_builddir", "/top_srcdir", NULL]

we're expecting to get "/top_srcdir" but g_build_filename is returning
"/top_builddir/top_srcdir". This results in gjs-unit looking for test
files in wrong directory and dying with

    ** ERROR:(gjs-unit.c:172):main: assertion failed: (dir != NULL)

Add custom function to handle absolute paths as we're expecting.

https://bugzilla.gnome.org/show_bug.cgi?id=613234