Export prototype in GJS_DEFINE_PROTO
When creating a new prototype using the GJS_DEFINE_PROTO macros,
save a static reference to the JSObject* representing the created
prototype. Previously only a reference to the JSClass structure was
saved, which is basically just a vtable without any actual state.
This is necessary to be able to create new instances of that prototype
without calling the provided constructor.

https://bugzilla.gnome.org/show_bug.cgi?id=614413