Massively clean up callback invocation
For callbacks, it greatly simplifies the code to make one fundamental
assumption:

  There is at most one (callback,user_data,GDestroyNotify) triple per function.

Multiple (scope call) callbacks are OK, but supporting say two callbacks
with GDestroyNotify is really painful implementation-wise, and I'm not
aware of a function with such a signature at the moment that lives in
the introspectable stack.  (GHashTable has one, but it doesn't count
being in GLib).

Once we have this first assumption, note that we can have a single
GDestroyNotify callback instance, and pass in for the "user_data"
what we need into that destroy notify.  This is another huge
simplification since there's no longer two different kinds of callbacks
(one for actual callbacks and one for GDestroyNotify).

Also, I made things more efficient; e.g. there's no longer a pointless
g_slice for the ffi_cif; we can just pack the cif into the larger
structure.

Finally, this patch makes passing "undefined" for a callback an
explicit error.  I can't think how that's useful.  You can still
pass "null" of course.

https://bugzilla.gnome.org/show_bug.cgi?id=615222