Initialize return value when a callback throws
If a callback throws an exception (or argument conversion fails), we
still need to return something to C. Currently the return value is
left uninitialized, and is junk memory.

Initialize the return value to 0/0.0/NULL depending on the type.
This might be an invalid return value or cause the caller to
segfault in certain cases, but we can't really do better.

https://bugzilla.gnome.org/show_bug.cgi?id=618201