[gjs] Add a byteArray module.
docs/gjs-byte-array.txt explains how to use it and its features.

The primary purpose of this for now will be to bind C APIs that use
byte buffers, such as read and write operations. As such, the
ByteArray class is pretty simple and is mostly just a blob you can
shovel between C APIs.

In the initial patch, ByteArray does not even support array operations
such as slice().

The enumeration support relies on defining each byte as a property
on the byte array object, which makes enumerating bytes in the array
even more inefficient than it already would be.

https://bugzilla.gnome.org/show_bug.cgi?id=571000