Handle wide ranging enum values better
Enums were being converted to and from jsval using JSVAL_TO_INT
and INT_TO_JSVAL. The latter is just wrong - INT_TO_JSVAL produces
junk results if called on an integer outside the range of
-0x40000000 - 0x3fffffff.

Just fixing that still leaves a problem - if we get 0x80000000 as
a flags value, and we treat it as (int)0x80000000 then we pass it
to another function that is using 'guint' for flags instead of the
enum type, we'll throw because the value is out of range.

So, this change goes a further and (by depending on gobject-introspection
changes in 629704) correctly maps all signed and unsigned 32 bit
constants onto the correct numeric value.

A new utility function gjs_value_to_int64() is added to convert from
jsval to gint64, since we are using int64 as an intermediate type for
holding enum values.

https://bugzilla.gnome.org/show_bug.cgi?id=629705