Cast users of JS_EnterLocalRootScope
In Xulrunner 1.9.3, this is now a no-op.  See
upstream bug: https://bugzilla.mozilla.org/show_bug.cgi?id=519949

Now in theory, we should be checking the return value and handling
that.  I wrote a nontrivial patch to do this, but decided it
wasn't worth it, since in newer xulrunners it's just pointless
code spaghetti, and in older, it's not like we're really
going to go OOM in reality.