Replace "load context" with a "import global"
The idea of the load context was to have a single context where modules
and native were defined and shared between different JSContexts in the
same runtime.

But this caused context switching when loading modules, which isn't very
tested in Spidermonkey and caused lots of use of gjs_move_exception()
and other complications.

Instead, switch to the idea of a "import global" object; this is the global
object that is used in the scope chain for modules, and is also where the
constructors for our native classes are stored, allowing native classes
to be shared between different contexts.

https://bugzilla.gnome.org/show_bug.cgi?id=622896