dbus: improve async callback handling
When using the "async" technqiue of exporting dbus methods, it's
the responsibility of the service to call the passed in callback
function when it's time to send the reply.

This callback function takes the return value or an array of return
values (matched in order one by one with the types mentioned in the
out signature for the method).

Some methods don't have any out signature though.  Right now, these
methods are required to call the callback with an empty array
argument.

This commit changes gjs to be more flexible and allow the caller to
invoke the callback with no arguments for this particular "empty
out signature" case.