boxed: Avoid recursion and conversion during construction
boxed_new previously called gjs_invoke_c_function_uncached() on
the first C constructor it found; this recursed and we would
wrap the boxed, creating a JSObject etc., only to unwrap it
and get the native pointer again.

For a reason I'm not going to debug in depth, this started failing
with xulrunner 2.  Regardless, we should avoid this insanity and
call the C function more directly with g_function_info_invoke,
which gives us the raw pointer we wanted anyways.

https://bugzilla.gnome.org/show_bug.cgi?id=632159