Throw an exception when failing to look up a foreign struct
Returning failure without setting an exception was causing control
to mysteriously jump out of the enclosing function. The function
is also changed to g_assert() rather than returning NULL when
a non-interface GITypeInfo is passed in - it's a case that should
not happen if the calling code is correct so throwing an exception
doesn't make sense.

https://bugzilla.gnome.org/show_bug.cgi?id=635359