Ignore empty elements in the search path
We were treating empty elements in the search path as meaning the
current directory. However, empty elements are easy to end up with
by mistakes in path construction, and treating them as the current
directory can result in extremely confusing results. If the current
directory is actually meant, it can be specified as '.'.

https://bugzilla.gnome.org/show_bug.cgi?id=635368