xulrunner2: Conditionally handle availability of JS_GetStringBytes
This upstream mozilla commit:
http://hg.mozilla.org/mozilla-central/changeset/f7171a41a816
removed JS_GetStringBytes in favor of an API which requires
the caller to allocate.  We were using this function in
a number of places, and most of those now need to malloc()
with the new API.  Rather than trying to use JS_GetStringBytes
as "const", and malloc() only with the new API, wrap it
and always malloc()/free().

Based on a patch originally from Sardem FF7 <sardemff7.pub@gmail.com>.

https://bugzilla.gnome.org/show_bug.cgi?id=635707