gjs_value_debug_string: Always return UTF-8
Returning whatever JS_GetStringBytes gave us will blow up if
the string contains non-UTF8 characters and we're trying to g_print
to a UTF-8 terminal (the standard case).  Since this is just a
debugging string, import a copy of _g_utf8_make_valid which
squashes it to UTF-8 in a useful way.