Check for a builtin module before beginning to load it
When used to check for builtin modules, import_native_file() would create a
module object, set properties on it, store it in the native class, fail
to load a builtin module of that name, and then back everything out.
Add a pre-check to avoid this. (It means that we do extract the module ID
twice when we actually load a built-in module, but built-in modules will
a small fraction of modules.)

V2: Modified by Colin Walters <walters@verbum.org> for the fact that
gjs_string_get_ascii() now mallocs.

https://bugzilla.gnome.org/show_bug.cgi?id=635371