Add rpath to gjs-internals-1.0.pc
Since gjs-internal-1.0.pc pulls in the mozilla-js PC file, anything
that uses it needs the same mozjs-not-in-system-directory workaround
that we already use for gjs-console. So, add the needed RPATH to
Libs. We need to use the GNU-specific -Wl,--rpath rather than the
libtool -R because we can't count on libtool always being used
to link the result of pkg-config --libs. Given the small number of
different linkers we'll encounter in practice, we can always add
configure checks to figure out the right flags if someone hits
problems.