GIRepositoryFunction: make it inherit from Function
Set the prototype of GIRepositoryFunction to Function.prototype, so
methods bind(), call() and apply() are available, and instanceof
reports correct results.
Also ensures that method invocations are applied on GObjects of the
correct type, and adds an explicit toString() method to GIRepositoryFunction,
since Function.prototype.toString() throws when called on objects
of the wrong class.

https://bugzilla.gnome.org/show_bug.cgi?id=646852