Complete porting to GDBus
Introduces a new introspected library, libgjs-gdbus, which hosts
GjsDBusImplementation, a standalone, signal-driven GDBusInterfaceSkeleton.
This is exposed, thanks to overrides, as Gio.DBusExportedObject, where
it gains some JS code to expose normal JS objects on the bus
and some glue to use E4X in place of XML strings for introspection.

https://bugzilla.gnome.org/show_bug.cgi?id=622921