Remove SIGSEGV, SIGABRT handlers
It's possible for backtrace() to itself invoke malloc() in at least
some versions of glibc.  Due to the fact that the malloc lock is not
recursive, this will in turn deadlock if we got SIGSEGV inside malloc().

Deadlocking is far worse than segfaulting, and most operating systems
nowadays have some sort of systemwide crash catching facility, so let's
just rely on that.