object: Support glib properties
To support installing custom properties, we must be able to have JS context and
JS object in set_property before even g_object_newv() returns. This is achieved
by using other internal construct properties that are set if the object is of a
custom type.

Use this so that we can implement a sane default implementation of
do_get_property/do_set_property that simply sets the corresponding field on the
JS object. Implementations that need more complex behavior can either use ES5
or Mozilla getters/setters, or implement their own override of do_get_property
and do_set_property.