Represent GTypes as a new boxed wrapper type
We can't natively represent a pointer as a regular value in JS (we have at
most 30-bit integers or doubles), so to store GTypes correctly, we need to
box them. Create a new native object that simply represents a GType, give
it a 'g_type_name' wrapper, and fix up the rest of the world so we no longer
treat a GType as a special integer.

https://bugzilla.gnome.org/show_bug.cgi?id=668427