Improve GType detection for unconstrained GValues
In ebc84492, we added a way to get a GType for a constructor.
Let's expand this to objects (by way of the JavaScript 'constructor'
property), and then use this mechanism to improve GType detection
in GValues. This can be used with APIs like GtkListStore that force
the user to create a GValue and expect the right type.

https://bugzilla.gnome.org/show_bug.cgi?id=669848