Reference GjsCallbackTrampoline during callbacks to JS
When we are calling back to Javascript, keep the GjsCallbackTrampoline
object alive for the duration of the call, since we need to use information
from it to process the return value even if the callback is removed during the
callback.

(This actually matters only for GI_SCOPE_TYPE_NOTIFIED, but we reference
in all cases for simplicity.)

See https://bugzilla.gnome.org/show_bug.cgi?id=643014#c15 for an example
of code triggering this problem.

https://bugzilla.gnome.org/show_bug.cgi?id=672003