Use a GTypeModule for registering custom types
This makes so that the types are installed as dynamic types, and
thus loaded and unloaded on demand. In particular, resources such
as signal class closures are now properly released when the use count
of these types drops to zero.
Additionally, to avoid unloading too early, each GObject prototype
has a now a reference on the GObjectClass. This means that the above
only applies at context destruction, and preserves previous semantics
wrt property and signal installation.

https://bugzilla.gnome.org/show_bug.cgi?id=679688