Add new GBytes API and conversions
GLib's annotations for GBytes have been updated so one can now do:

GLib.Bytes.new([1,2,3])

This allows us to construct them natively, which is nice.  But we
should go further.

Now, we should have the ability to convert them to a JS array.  In
gjs, we have the byteArray class which "looks like" a JS array, but
actually points to a native array, since it's a lot more efficient.

So, change the internal byteArray class so that it supports holding
a GBytes *or* a GByteArray.  It's used as a container as well as
a bridge from JS -> native conversions for the overrides code.

Add an override method "toArray" that allows converting a GBytes to a
"gjs byteArray", which in turn has automatic conversions to
e.g. guint8+len pair as used from C.

https://bugzilla.gnome.org/show_bug.cgi?id=685431