ABI CHANGE: Switch String -> uint8 array coercion to be UTF8
Previously for historical reasons, we treated JavaScript Strings as
byte array containers when in the context of a uint8 array conversion.
This is not what "most" users expect, particularly when taking UTF-8
data from one GLib function and passing it to a different GLib
function that accepts a uint8 array.

https://bugzilla.gnome.org/show_bug.cgi?id=680730