Remove obsolete handling of binary data and move to CStringsAreUTF8
All C strings are UTF8 in the glib world, so that setting make sense
for gjs, and the reason it was previously off was to allow a broken
way to represent binary data.
Now that we don't do that any more, we can remove a lot of useless code
and just let Spidermonkey do the right thing.

https://bugzilla.gnome.org/show_bug.cgi?id=680730