Use GLib for JS logging
Reserve our custom logging framework for debug logging, and use standard
g_log functions for log() and logError().
This allows us to finally have readable error messages, it associates
them with a program and process when necessary, and it prepares us
for a future where glib logs to the journal by default.

https://bugzilla.gnome.org/show_bug.cgi?id=690984