byteArray: Remove prototype checks from func implementations
As we don't use JSCLASS_CONSTRUCT_PROTOTYPE anymore, we don't
have a private on the prototype, which means that all instances
of a NULL priv are the prototype. As far as I can tell, the
"wrong class" case is impossible, as all of these functions
are only installed on an object with a JSClass of ByteArray.

https://bugzilla.gnome.org/show_bug.cgi?id=690982