context: Don't allow users to pass external runtimes to GjsContext
Upstream wants to merge JSContext and JSRuntime, because the distinction
has been a constant source and confusion and pain as they rearrange the
runtime. As a first step towards making this shift on the gjs side, we
need to discourage external consumers of gjs from making the distinction.

Remove support for the "runtime" property which allows consumers of gjs
to specify their own JS runtime.

https://bugzilla.gnome.org/show_bug.cgi?id=691109