examples: fix http-server.js example
For some reason, this example was reimplementing a bunch of
SoupServer's functionality itself in JavaScript, rather than using
SoupServer in the expected way. (This may be because either gi or
Soup-2.4.gir was not up to the task at the time the example was
written.)

Anyway, fix. Also, spice it up a bit.

https://bugzilla.gnome.org/show_bug.cgi?id=694300