context: lock access to GC idle
There's a small race where the GC notification idle id might get
zeroed before the idle runs, if the GC callback is run from a
non-main thread.

That means there's the potential for the GC notification idle to
get dispatched after the context is finalized.

This commit fixes that by protecting access to the idle id with
a lock.

https://bugzilla.gnome.org/show_bug.cgi?id=670200