Expose the launched binary to JS
Usually in various languages the launched program is available
as ARGV[0], but in the gjs case, ARGV only contains command line
arguments, so we need a separate property to retrieve the launched
script.
This is now exposed as System.programInvocationName.
The use case is allowing package.js to determine if the app is
launched from the source without hacks.

https://bugzilla.gnome.org/show_bug.cgi?id=699784