Stop walking the stack manually
In JS 17, and when dealing with optimized frames without "arguments",
our technique to build the stack can fail and throw an exception.
Normally we would override it when we set the real exception, but
in the GDBus test suite we didn't, so there was always this
unnoticed "Debugger scope is not live" error, until it was made
fatal by glib changes.

https://bugzilla.gnome.org/show_bug.cgi?id=690984