Fix regression retriving GTypes from instances
The GValue marshalling code calls get_actual_gtype() on instances.
To handle that, we need three levels of recursion: from the
instance to the constructor, from the constructor to the GType
object and finally from GType object to the actual GType.

This regressed in df28a9c070459cc90b0ea0db19bf8e0568ac396a,
which was later reverted, but that left the original bug unfixed.

https://bugzilla.gnome.org/show_bug.cgi?id=699878