Implement GTypeClass and GTypeInterface arguments
It is common in the GObject world to have non trivial functionality
exposed as GObjectClass/GTypeClass methods, to be used from
class_init. One example is the GtkWidget template system.
This commit implements marshalling of GType structures as
JS classes (constructors), going through the associated GType.

Note that this commit breaks the testGObjectClass.js test

https://bugzilla.gnome.org/show_bug.cgi?id=700347