tests: Add new installed-tests/scripts test mechanism
The unit tests are not suitable for all sorts of tests; in particular,
I can't test System.exit() from there since it loads all of the
tests into one process.

So this patch adds a new type of "scripts" test which is literally
just a gjs script; if it exits successfully, the test passed.

https://bugzilla.gnome.org/show_bug.cgi?id=703826