tests: Disable locale conversion failure tests
This is for the mozjs24 port.

At the moment, mozjs24 must be built with --disable-intl-api for the
locale callbacks to work.  In theory the locale callbacks are no
longer needed since this is handled by ICU, however it looks like js24
will ship with an in-tree ICU, which many distros might disable.

https://bugzilla.gnome.org/show_bug.cgi?id=711046