Added GjsCoverage object (and thus code coverage support).
This object effectively just takes a connection to the single step
handler and records how many times we hit each line. It then provides
a function to write this data out to a GFile, or if the GFile provided is
NULL, writes to $(file).js.info.

To prevent clutter, paths considered to be part of the coverage report
must be specified inclusively, anything not part of those paths will not
be included in the coverage report.

The coverage tool is integrated in to gjs-console, you can use it by
specifying -C or --coverage-path to add a path to generate coverage reports
for and using --coverage-output to write coverage data to a single file.