jsapi-util: Add a gjs_parse_call_args variant, that takes JS::CallArgs
JS::CallArgs is now a recommended replacement for the argc/vp pair, and
in the future, JSNative will take a JS::CallArgs directly, instead of a
argc/vp pair. For now, use the semi-deprecated "array" member to share
the same code as the old argc/vp interface.

When upstream deprecates argc/vp, we'll port everything to JS::CallArgs.

This is because I spend almost half an hour finding a bug with the old
interface (I called JS_SET_RVAL on the argv instead of vp), and they
recommended that I use JS::CallArgs instead.