object toggle notify: do more work without deferring to an idle
The only case we really need an idle is when called from a secondary
thread (and even there it's arguable the idle is fixing anything).
Inside the main thread we're in full control of what the GC is
doing, and we know when it's safe and when it's not to touch the
JSAPI. Deferring to an idle while the GC is already in the sweeping
phase is late, by the time the idle runs the JS object is already
dead and we're accessing freed memory.

https://bugzilla.gnome.org/show_bug.cgi?id=725024