GjsDBusImplementation: unref the invocation passed to method_call()
In C code, the reference passed to the method_call() virtual function
would be passed to the return_value/error function that is eventually
called. But since these functions are called from JS, and the calling
convention is normalized, we need to unref reference passed to
method_call ourselves. See GLib bug 738259 for a more complete discussion.

https://bugzilla.gnome.org/show_bug.cgi?id=738122