promise: Get node-isms to work
This adds some stubs to lie.js for node.js-related objects that it
expects to be present, such as "process".

It implements "require" as well, and provides an implementation of
immediate.js using GLib. It would have been somewhat feasible to port
immediate.js, but it requires either a DOM or setTimeout(), so we would
have had to basically write this implementation anyway.

It also wraps the whole thing in an IIFE so that symbols are not leaked
out of the scope, and returns the Promise object so that it will be the
result of calling JS::Evaluate() on the code.

https://bugzilla.gnome.org/show_bug.cgi?id=608450